some refactoring, surfaces+emission switches + bugfixes
