+++ tif2blender/load.py
@@ -5,41 +5,12 @@ from bpy.props import (StringProperty, FloatProperty,
 from pathlib import Path
 import numpy as np
 
-from .initial_global_settings import preset_environment
+from .initial_global_settings import preset_environment, preset_em_environment
 from .handle_blender_structs import *
 from .load_components import *
+from .unpack_tif import unpack_tif, changePathTif
 
 from mathutils import Matrix
-
-
-
-def changePathTif(self, context):
-    # infers metadata, resets to default if not found
-    # the raise gets handled upstream, so only prints to cli, somehow.
-    try: 
-        import tifffile
-        with tifffile.TiffFile(context.scene.T2B_input_file) as ifstif:
-            try:
-                context.scene.T2B_axes_order = ifstif.series[0].axes.lower().replace('s', 'c')
-            except Exception as e:
-                print(e)
-                context.scene.property_unset("T2B_axes_order")
-            try:
-                context.scene.T2B_xy_size = ifstif.pages[0].tags['XResolution'].value[1]/ifstif.pages[0].tags['XResolution'].value[0]
-            except Exception as e:
-                print(e)
-                context.scene.property_unset("T2B_xy_size")
-            try:
-                context.scene.T2B_z_size = dict(ifstif.imagej_metadata)['spacing']
-            except Exception as e:
-                print(e)
-                context.scene.property_unset("T2B_z_size")
-    except Exception as e:
-        context.scene.property_unset("axes_order")
-        context.scene.property_unset("xy_size")
-        context.scene.property_unset("z_size")
-        raise
-    return
 
 bpy.types.Scene.T2B_remake = bpy.props.BoolProperty(
     name = "TL_remake", 
@@ -50,6 +21,18 @@ bpy.types.Scene.T2B_remake = bpy.props.BoolProperty(
 bpy.types.Scene.T2B_preset_environment = bpy.props.BoolProperty(
     name = "TL_preset_environment", 
     description = "Set environment variables",
+    default = True
+    )
+
+bpy.types.Scene.T2B_Emission = bpy.props.BoolProperty(
+    name = "TL_EM", 
+    description = "Electron microscopy data preset (absorbent volume)",
+    default = True
+    )
+
+bpy.types.Scene.T2B_Surface = bpy.props.BoolProperty(
+    name = "TL_EM", 
+    description = "Load isosurface object",
     default = True
     )
 
@@ -85,118 +68,28 @@ bpy.types.Scene.T2B_z_size = FloatProperty(
         description="z physical pixel size in micrometer",
         default=1.0)
 
+
 bpy.types.Scene.T2B_mask_channels = StringProperty(
         name="",
         description="channels with an integer label mask",
         )
 
-
-def unpack_tif(input_file, axes_order, test=False):
-    import tifffile
-    with tifffile.TiffFile(input_file) as ifstif:
-        imgdata = ifstif.asarray()
-
-    if len(axes_order) != len(imgdata.shape):
-        raise ValueError("axes_order length does not match data shape: " + str(imgdata.shape))
-
-    for dim in 'tzcyx':
-        if dim not in axes_order:
-            imgdata = np.expand_dims(imgdata,axis=0)
-            axes_order = dim + axes_order
-
-    mask_channels = []
-    if bpy.context.scene.T2B_mask_channels != '':
-        try:
-            mask_channels = [int(ch.strip()) for ch in bpy.context.scene.T2B_mask_channels.split(',') if '-' not in ch]
-            # mask_channels.extend([list(np.arange(int(ch.strip().split('-')[0]),int(ch.strip().split('-')[1]))) for ch in bpy.context.scene.T2B_mask_channels.split(',')if '-' in ch])
-        except:
-            raise ValueError("could not interpret maskchannels")
-        if max(mask_channels) >= imgdata.shape[axes_order.find('c')]:
-            raise ValueError(f"mask channel is too high, max is {imgdata.shape[axes_order.find('c')]-1}, it starts counting at 0" )
-    
-    mask_arrays = {}
-    for ch in mask_channels:
-        mask_arrays[ch] = imgdata.take(indices=ch, axis=axes_order.find('c'))
-        # maybe check input here?
-    # volume_array = np.delete(imgdata, mask_channels, axis=axes_order.find('c'))
-    volume_array = imgdata
-    
-    # normalize values per channel
-    volume_array = volume_array.astype(np.float32)
-    ch_first = np.moveaxis(volume_array, axes_order.find('c'), 0)
-    for chix, chdata in enumerate(ch_first):
-        ch_first[chix] -= np.min(chdata)
-        ch_first[chix] /= np.max(chdata)
-    channels = volume_array.shape[axes_order.find('c')] - len(mask_channels)
-    all_channels = imgdata.shape[axes_order.find('c')]
-    
-
-    # otsu compute in z MIP
-    otsus = [-1] * all_channels
-    
-    
-    for channel in range(channels):
-        ix = sorted(list(set(range(all_channels)) - set(mask_channels)))[channel]
-        if channels > 1:
-            im = volume_array.take(indices=channel, axis=axes_order.find('c'))
-        else:
-            im = volume_array
-        ch_axes = axes_order.replace("c","")
-        z_MIP = np.amax(im, axis = ch_axes.find('z'))
-        threshold_range = np.linspace(0,1,101)
-        try:
-            from skimage.filters import threshold_otsu
-            otsus[ix] = threshold_otsu(z_MIP)
-            print('used scikit image' )
-        except:
-            criterias = [compute_otsu_criteria(z_MIP, th) for th in threshold_range]
-            otsus[ix] = threshold_range[np.argmin(criterias)]
-
-    size_px = np.array([imgdata.shape[axes_order.find('x')], imgdata.shape[axes_order.find('y')], imgdata.shape[axes_order.find('z')]])
-    print(otsus)
-    return volume_array, mask_arrays, otsus, size_px, axes_order
-
-
-# adapted from https://en.wikipedia.org/wiki/Otsu%27s_method
-# TODO see if skimage is there and then use that one + do some planes, and not MIP
-def compute_otsu_criteria(im, th):
-    """Otsu's method to compute criteria."""
-    thresholded_im = np.zeros(im.shape)
-    thresholded_im[im >= th] = 1
-
-    # compute weights
-    nb_pixels = im.size
-    nb_pixels1 = np.count_nonzero(thresholded_im)
-    weight1 = nb_pixels1 / nb_pixels
-    weight0 = 1 - weight1
-
-    # if one of the classes is empty, eg all pixels are below or above the threshold, that threshold will not be considered
-    # in the search for the best threshold
-    if weight1 == 0 or weight0 == 0:
-        return np.inf
-
-    # find all pixels belonging to each class
-    val_pixels1 = im[thresholded_im == 1]
-    val_pixels0 = im[thresholded_im == 0]
-
-    # compute variance of these classes
-    var1 = np.var(val_pixels1) if len(val_pixels1) > 0 else 0
-    var0 = np.var(val_pixels0) if len(val_pixels0) > 0 else 0
-
-    return weight0 * var0 + weight1 * var1
-
-
 def load():
-    # Handle all globals to be scoped after this
+    # Handle all globals to be scoped after this - axes order in particular gets internally changed sometimes 
     input_file = bpy.context.scene.T2B_input_file
     axes_order = bpy.context.scene.T2B_axes_order
     remake = bpy.context.scene.T2B_remake
     xy_size = bpy.context.scene.T2B_xy_size
     z_size = bpy.context.scene.T2B_z_size
     cache_dir = Path(bpy.context.scene.T2B_cache_dir) / Path(input_file).stem
+    mask_channels = bpy.context.scene.T2B_mask_channels
+    surfaces = bpy.context.scene.T2B_Surface
+    emission = bpy.context.scene.T2B_Emission
 
     if bpy.context.scene.T2B_preset_environment:
         preset_environment()    
+        if not emission:
+            preset_em_environment()
     
     if xy_size <= 0  or z_size <= 0:
         raise ValueError("cannot do zero-size pixels")
@@ -209,7 +102,7 @@ def load():
     cache_coll = collection_by_name(Path(input_file).stem, supercollections=['cache'], duplicate=True)
     
     # pads axes order with 1-size elements for missing axes
-    volume_array, mask_arrays, otsus, size_px, axes_order = unpack_tif(input_file, axes_order)
+    volume_array, mask_arrays, otsus, size_px, axes_order = unpack_tif(input_file, axes_order, mask_channels)
 
     to_be_parented = []
 
@@ -220,10 +113,12 @@ def load():
 
     if len(mask_arrays) != len(otsus):
         vdb_files, bbox_px = array_to_vdb_files(volume_array, axes_order, remake, cache_dir)
-        vol_obj, vol_coll = load_volume(vdb_files, bbox_px, otsus, scale, cache_coll, base_coll)
-        surf_obj = load_surfaces(vol_coll, otsus, scale, cache_coll, base_coll)
+        vol_obj, vol_coll = load_volume(vdb_files, bbox_px, otsus, scale, cache_coll, base_coll, emission)
         to_be_parented.extend([vol for vol in vol_coll.all_objects])
-        to_be_parented.extend([vol_obj, surf_obj])
+        to_be_parented.extend([vol_obj])
+        if surfaces:
+            surf_obj = load_surfaces(vol_coll, otsus, scale, cache_coll, base_coll)
+            to_be_parented.extend([surf_obj])
     
     if len(mask_arrays) > 0:
         mask_obj, mask_colls = load_labelmask(mask_arrays, scale, cache_coll, base_coll, cache_dir, remake, axes_order)
