+++ tif2blender/t2b_nodes/nodeBoundedMapRange.py
@@ -1,5 +1,7 @@
 import bpy 
 
+# UNUSED currently
+
 # This is a reskin of Map Range node to make it easier to understand for novice users
 # by removing things unnecessary for microscopy data, and bounding values to tif2blender
 # ranges
